requirements.txt
discordwebhook==1.0.3
# kafka-python-ng==2.2.3
psycopg==3.2.3
psycopg2-binary==2.9.10
pip==24.0
SQLAlchemy==2.0.36
jinja2==3.1.2
aiohttp==3.9.1
pydantic-settings==2.1.0
aiokafka==0.8.0
loguru==0.7.2
apprise==1.7.3
